+++ tests/unit_tests/model/test_orders.py
@@ -1454,7 +1454,7 @@ class TestOrders:
         assert bracket.orders[2].parent_order_id == ClientOrderId("O-19700101-000000-000-001-1")
         assert bracket.ts_init == 0
 
-    def test_bracket_stop_limit_entry_stop_limit_tp_order_list(self):
+    def test_bracket_limit_if_touched_entry_stop_limit_tp_order_list(self):
         # Arrange, Act
         bracket = self.order_factory.bracket(
             AUDUSD_SIM.id,
@@ -1491,6 +1491,63 @@ class TestOrders:
         assert bracket.orders[1].quantity == Quantity.from_int(100_000)
         assert bracket.orders[2].quantity == Quantity.from_int(100_000)
         assert bracket.orders[1].trigger_price == Price.from_str("0.99990")
+        assert bracket.orders[2].price == Price.from_str("1.00010")
+        assert bracket.orders[1].time_in_force == TimeInForce.GTC
+        assert bracket.orders[2].time_in_force == TimeInForce.GTC
+        assert bracket.orders[1].expire_time is None
+        assert bracket.orders[2].expire_time is None
+        assert bracket.orders[0].contingency_type == ContingencyType.OTO
+        assert bracket.orders[1].contingency_type == ContingencyType.OUO
+        assert bracket.orders[2].contingency_type == ContingencyType.OUO
+        assert bracket.orders[0].linked_order_ids == [
+            ClientOrderId("O-19700101-000000-000-001-2"),
+            ClientOrderId("O-19700101-000000-000-001-3"),
+        ]
+        assert bracket.orders[1].linked_order_ids == [ClientOrderId("O-19700101-000000-000-001-3")]
+        assert bracket.orders[2].linked_order_ids == [ClientOrderId("O-19700101-000000-000-001-2")]
+        assert bracket.orders[1].parent_order_id == ClientOrderId("O-19700101-000000-000-001-1")
+        assert bracket.orders[2].parent_order_id == ClientOrderId("O-19700101-000000-000-001-1")
+        assert bracket.ts_init == 0
+
+    def test_bracket_stop_limit_entry_stop_limit_tp_order_list(self):
+        # Arrange, Act
+        bracket = self.order_factory.bracket(
+            AUDUSD_SIM.id,
+            OrderSide.BUY,
+            Quantity.from_int(100_000),
+            entry_trigger_price=Price.from_str("1.00000"),
+            entry_price=Price.from_str("1.00001"),
+            sl_trigger_price=Price.from_str("0.99990"),
+            tp_trigger_price=Price.from_str("1.00010"),
+            tp_price=Price.from_str("1.00010"),
+            time_in_force=TimeInForce.GTC,
+            entry_order_type=OrderType.STOP_LIMIT,
+            tp_order_type=OrderType.LIMIT_IF_TOUCHED,
+        )
+
+        # Assert
+        assert bracket.id == OrderListId("OL-19700101-000000-000-001-1")
+        assert bracket.instrument_id == AUDUSD_SIM.id
+        assert len(bracket) == 3
+        assert len(bracket.orders) == 3
+        assert bracket.orders[0].order_type == OrderType.STOP_LIMIT
+        assert bracket.orders[1].order_type == OrderType.STOP_MARKET
+        assert bracket.orders[2].order_type == OrderType.LIMIT_IF_TOUCHED
+        assert bracket.orders[0].instrument_id == AUDUSD_SIM.id
+        assert bracket.orders[1].instrument_id == AUDUSD_SIM.id
+        assert bracket.orders[2].instrument_id == AUDUSD_SIM.id
+        assert bracket.orders[0].client_order_id == ClientOrderId("O-19700101-000000-000-001-1")
+        assert bracket.orders[1].client_order_id == ClientOrderId("O-19700101-000000-000-001-2")
+        assert bracket.orders[2].client_order_id == ClientOrderId("O-19700101-000000-000-001-3")
+        assert bracket.orders[0].side == OrderSide.BUY
+        assert bracket.orders[1].side == OrderSide.SELL
+        assert bracket.orders[2].side == OrderSide.SELL
+        assert bracket.orders[0].quantity == Quantity.from_int(100_000)
+        assert bracket.orders[1].quantity == Quantity.from_int(100_000)
+        assert bracket.orders[2].quantity == Quantity.from_int(100_000)
+        assert bracket.orders[0].trigger_price == Price.from_str("1.00000")
+        assert bracket.orders[1].trigger_price == Price.from_str("0.99990")
+        assert bracket.orders[0].price == Price.from_str("1.00001")
         assert bracket.orders[2].price == Price.from_str("1.00010")
         assert bracket.orders[1].time_in_force == TimeInForce.GTC
         assert bracket.orders[2].time_in_force == TimeInForce.GTC